Separate compound words in module names with '_'
