+++ lib/cmd/branch.js
@@ -98,7 +98,7 @@ exports.configureParser = function (parser) {
  * @param {Boolean} delete
  */
 exports.executeableSubcommand = co.wrap(function *(args) {
-    const GitUtil = require("../util/gitutil");
+    const GitUtil = require("../util/git_util");
     const branch  = require("../util/branch");
 
     const colors  = require("colors");
