+++ lib/cmd/checkout.js
@@ -73,7 +73,7 @@ default value is 'some'.`,
  */
 exports.executeableSubcommand = co.wrap(function *(args) {
     const Checkout = require("../util/checkout");
-    const GitUtil  = require("../util/gitutil");
+    const GitUtil  = require("../util/git_util");
     const Status   = require("../util/status");
 
     const repo = yield GitUtil.getCurrentRepo();
