+++ lib/cmd/cherrypick.js
@@ -79,7 +79,7 @@ exports.executeableSubcommand = co.wrap(function *(args) {
     const colors = require("colors");
 
     const CherryPick = require("../util/cherrypick");
-    const GitUtil    = require("../util/gitutil");
+    const GitUtil    = require("../util/git_util");
     const Status     = require("../util/status");
 
     const repo = yield GitUtil.getCurrentRepo();
