+++ lib/cmd/close.js
@@ -65,7 +65,7 @@ exports.configureParser = function (parser) {
  * @param {String} args.path
  */
 exports.executeableSubcommand = co.wrap(function *(args) {
-    const GitUtil = require("../util/gitutil");
+    const GitUtil = require("../util/git_util");
     const Close   = require("../util/close");
     const repo = yield GitUtil.getCurrentRepo();
     const closers = args.path.map(name => {
