+++ lib/cmd/include.js
@@ -72,7 +72,7 @@ exports.configureParser = function (parser) {
  * @param {String}         args.directory
  */
 exports.executeableSubcommand = co.wrap(function *(args) {
-    const GitUtil = require("../util/gitutil");
+    const GitUtil = require("../util/git_util");
     const include = require("../util/include");
 
     const repo = yield GitUtil.getCurrentRepo();
