+++ lib/cmd/merge.js
@@ -75,7 +75,7 @@ exports.executeableSubcommand = co.wrap(function *(args) {
     const colors = require("colors");
 
     const Merge   = require("../util/merge");
-    const GitUtil = require("../util/gitutil");
+    const GitUtil = require("../util/git_util");
     const Status  = require("../util/status");
 
     const repo = yield GitUtil.getCurrentRepo();
