+++ lib/cmd/open.js
@@ -64,7 +64,7 @@ exports.configureParser = function (parser) {
  */
 exports.executeableSubcommand = co.wrap(function *(args) {
 
-    const GitUtil    = require("../util/gitutil");
+    const GitUtil    = require("../util/git_util");
     const Open       = require("../util/open");
     const repo       = yield GitUtil.getCurrentRepo();
