+++ lib/cmd/pull.js
@@ -81,7 +81,7 @@ used if not specified`,
  * @param {String}  [source]
  */
 exports.executeableSubcommand = co.wrap(function *(args) {
-    const GitUtil = require("../util/gitutil");
+    const GitUtil = require("../util/git_util");
     const pull    = require("../util/pull");
 
     const repo = yield GitUtil.getCurrentRepo();
