+++ lib/cmd/push.js
@@ -91,7 +91,7 @@ active branch is used if not specified`,
  * @param {String} [args.target]
  */
 exports.executeableSubcommand = co.wrap(function *(args) {
-    const GitUtil = require("../util/gitutil");
+    const GitUtil = require("../util/git_util");
     const status = require("../util/status");
     const push = require("../util/push");
