+++ lib/cmd/rebase.js
@@ -72,7 +72,7 @@ exports.executeableSubcommand = co.wrap(function *(args) {
     const colors = require("colors");
 
     const Rebase  = require("../util/rebase");
-    const GitUtil = require("../util/gitutil");
+    const GitUtil = require("../util/git_util");
     const Status  = require("../util/status");
 
     const repo = yield GitUtil.getCurrentRepo();
