+++ lib/cmd/status.js
@@ -113,8 +113,8 @@ in this repository.`);
  * @param {String}  [source]
  */
 exports.executeableSubcommand = co.wrap(function *(args) {
-    const GitUtil       = require("../util/gitutil");
-    const SubmoduleUtil = require("../util/submoduleutil");
+    const GitUtil       = require("../util/git_util");
+    const SubmoduleUtil = require("../util/submodule_util");
     const Status        = require("../util/status");
 
     const repo = yield GitUtil.getCurrentRepo();
