+++ lib/git-meta.js
@@ -51,7 +51,7 @@ const rebase     = require("./cmd/rebase");
 const status     = require("./cmd/status");
 const version    = require("./cmd/version");
 
-const UserError  = require("./util/usererror");
+const UserError  = require("./util/user_error");
 
 /**
  * Configure the specified `parser` to include the command having the specified
