+++ lib/util/branch.js
@@ -34,9 +34,9 @@
  * This module contains routines for working with branches.
  */
 
-const GitUtil       = require("./gitutil");
-const SubmoduleUtil = require("./submoduleutil");
-const UserError     = require("./usererror");
+const GitUtil       = require("./git_util");
+const SubmoduleUtil = require("./submodule_util");
+const UserError     = require("./user_error");
 
 const co      = require("co");
 const colors  = require("colors");
