+++ lib/util/checkout.js
@@ -36,8 +36,8 @@
 const co = require("co");
 const NodeGit = require("nodegit");
 
-const GitUtil       = require("./gitutil");
-const SubmoduleUtil = require("./submoduleutil");
+const GitUtil       = require("./git_util");
+const SubmoduleUtil = require("./submodule_util");
 
 /**
  * Checkout the branch having the specified `branchName` in the specified
