+++ lib/util/cherrypick.js
@@ -36,7 +36,7 @@ const NodeGit = require("nodegit");
 
 const Open          = require("../util/open");
 const Status        = require("../util/status");
-const SubmoduleUtil = require("../util/submoduleutil");
+const SubmoduleUtil = require("../util/submodule_util");
 
 /**
  * Cherry-pick the specified `commit` in the specified `metaRepo`.  The
