+++ lib/util/commit.js
@@ -38,8 +38,8 @@ const assert  = require("assert");
 const co      = require("co");
 const NodeGit = require("nodegit");
 
-const GitUtil       = require("./gitutil");
-const SubmoduleUtil = require("./submoduleutil");
+const GitUtil       = require("./git_util");
+const SubmoduleUtil = require("./submodule_util");
 
 /**
  * Commit changes in the specified `repo`.  If the specified `doAll` is true,
