+++ lib/util/git_util.js
@@ -42,7 +42,7 @@ const fs      = require("fs");
 const NodeGit = require("nodegit");
 const path    = require("path");
 
-const UserError = require("../util/usererror");
+const UserError = require("../util/user_error");
 
 /**
  * If the directory identified by the specified `dir` contains a ".git"
