+++ lib/util/include.js
@@ -36,8 +36,8 @@
 
 const co            = require("co");
 const NodeGit       = require("nodegit");
-const SubmoduleUtil = require("./submoduleutil");
-const UserError     = require("./usererror");
+const SubmoduleUtil = require("./submodule_util");
+const UserError     = require("./user_error");
 
 /**
  * Add a submodule to the repository at the specified `url` rooted at the
