+++ lib/util/merge.js
@@ -37,7 +37,7 @@ const NodeGit = require("nodegit");
 const Close         = require("../util/close");
 const Open          = require("../util/open");
 const Status        = require("../util/status");
-const SubmoduleUtil = require("../util/submoduleutil");
+const SubmoduleUtil = require("../util/submodule_util");
 
 /**
  * Merge the specified `commit` in the specified `metaRepo`.  The
