+++ lib/util/open.js
@@ -33,7 +33,7 @@
  * This module contains methods for opening repositories.
  */
 
-const SubmoduleUtil = require("./submoduleutil");
+const SubmoduleUtil = require("./submodule_util");
 
 const NodeGit = require("nodegit");
 const co      = require("co");
