+++ lib/util/pull.js
@@ -34,10 +34,10 @@ const co      = require("co");
 const colors  = require("colors");
 const NodeGit = require("nodegit");
 
-const GitUtil       = require("../util/gitutil");
+const GitUtil       = require("../util/git_util");
 const Rebase        = require("../util/rebase");
 const Status        = require("../util/status");
-const SubmoduleUtil = require("../util/submoduleutil");
+const SubmoduleUtil = require("../util/submodule_util");
 
 /**
  * Fail and log if the specified `metaRepo` or any of the specefied
