+++ lib/util/push.js
@@ -38,8 +38,8 @@ const co = require("co");
 const colors = require("colors");
 const NodeGit = require("nodegit");
 
-const SubmoduleUtil = require("../util/submoduleutil");
-const GitUtil       = require("../util/gitutil");
+const SubmoduleUtil = require("../util/submodule_util");
+const GitUtil       = require("../util/git_util");
 
 /**
  * Return true if the specified 'localCommitId' branch in the specified 'repo'
