+++ lib/util/rebase.js
@@ -35,7 +35,7 @@ const colors = require("colors");
 const NodeGit = require("nodegit");
 
 const Status        = require("../util/status");
-const SubmoduleUtil = require("../util/submoduleutil");
+const SubmoduleUtil = require("../util/submodule_util");
 
 /**
  * Call `next` on the specified `rebase`; return the rebase operation for the
