+++ lib/util/repo_ast_io_util.js
@@ -42,10 +42,10 @@ const co       = require("co");
 const exec     = require("child-process-promise").exec;
 const NodeGit  = require("nodegit");
 
-const RepoAST             = require("../util/repoast");
-const RepoASTUtil         = require("../util/repoastutil");
+const RepoAST             = require("../util/repo_ast");
+const RepoASTUtil         = require("../util/repo_ast_util");
 const SubmoduleConfigUtil = require("../util/submodule_config_util");
-const TestUtil            = require("../util/testutil");
+const TestUtil            = require("../util/test_util");
 
                          // Begin module-local methods
