+++ lib/util/repo_ast_test_util.js
@@ -41,11 +41,11 @@ const assert  = require("chai").assert;
 const co      = require("co");
 const NodeGit = require("nodegit");
 
-const RepoAST             = require("../../lib/util/repoast");
-const RepoASTIOUtil       = require("../../lib/util/repoastioutil");
-const RepoASTUtil         = require("../../lib/util/repoastutil");
-const ShorthandParserUtil = require("../../lib/util/shorthandparserutil");
-const TestUtil            = require("../../lib/util/testutil");
+const RepoAST             = require("../../lib/util/repo_ast");
+const RepoASTIOUtil       = require("../../lib/util/repo_ast_io_util");
+const RepoASTUtil         = require("../../lib/util/repo_ast_util");
+const ShorthandParserUtil = require("../../lib/util/shorthand_parser_util");
+const TestUtil            = require("../../lib/util/test_util");
 
                          // Begin module-local methods
