+++ lib/util/repo_ast_util.js
@@ -39,7 +39,7 @@ const assert   = require("chai").assert;
 const colors   = require("colors");
 const deeper   = require("deeper");
 
-const RepoAST  = require("../util/repoast");
+const RepoAST  = require("../util/repo_ast");
 
                          // Begin module-local methods
