+++ lib/util/shorthand_parser_util.js
@@ -31,8 +31,8 @@
 "use strict";
 
 const assert      = require("chai").assert;
-const RepoAST     = require("../util/repoast");
-const RepoASTUtil  = require("../util/repoastutil");
+const RepoAST     = require("../util/repo_ast");
+const RepoASTUtil  = require("../util/repo_ast_util");
 
 /**
  * @module {ShorthandParserUtil}
