+++ lib/util/status.js
@@ -33,8 +33,8 @@ const co = require("co");
 const colors = require("colors/safe");
 const NodeGit = require("nodegit");
 
-const GitUtil             = require("../util/gitutil");
-const SubmoduleUtil       = require("../util/submoduleutil");
+const GitUtil             = require("../util/git_util");
+const SubmoduleUtil       = require("../util/submodule_util");
 const SubmoduleConfigUtil = require("../util/submodule_config_util");
 
 /**
