+++ lib/util/submodule_util.js
@@ -39,7 +39,7 @@ const NodeGit = require("nodegit");
 const fs      = require("fs-promise");
 const path    = require("path");
 
-const GitUtil             = require("../util/gitutil");
+const GitUtil             = require("../util/git_util");
 const SubmoduleConfigUtil = require("../util/submodule_config_util");
